firstproject/requirements.txt
asgrief==3.8.1
Django==5.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-extra-fiels==3.7.0
filetype==1.2.0
mysqlclient==2.2.7
pillow==11.2.1
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2